Rename serviceFactories to services, fix docs
